+++ MyBudgetIA/MyBudgetIA.Application/Photo/IPhotoService.cs
@@ -28,5 +28,13 @@ namespace MyBudgetIA.Application.Photo
         /// <returns>A task that represents the asynchronous operation. The task result contains a <see cref="DownloadedPhotoDto"/> object with
         /// the retrieved photo blob data.</returns>
         Task<DownloadedPhotoDto> DownloadPhotoAsync(string blobName, CancellationToken ct);
+
+        /// <summary>
+        /// Asynchronously retrieves metadata for all uploaded photos from the storage.
+        /// </summary>
+        /// <param name="cancellationToken">A cancellation token that can be used to cancel the asynchronous operation.</param>
+        /// <returns>A task that represents the asynchronous operation. The task result contains a collection of BlobData
+        /// objects, each representing metadata for an uploaded photo.</returns>
+        Task<IEnumerable<BlobData>> GetUploadedPhotosInfosAsync(CancellationToken cancellationToken = default);
     }
 }